requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.5
django-ajax-selects==1.7.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-widget-tweaks==1.4.3
gspread==3.1.0
gunicorn==19.9.0
httplib2==0.12.0
idna==2.8
oauth2client==4.1.3
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
